pyproject.toml
[project]
name = "nyc-yellow-taxi-public-data-analyzer"
version = "0.1.0"
description = "Production-style ELT data platform for analyzing NYC Yellow Taxi trip data"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.83",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.4.0",
    "black>=24.0.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "boto3-stubs[s3]>=1.42.83",
]

[tool.ruff]
target-version = "py313"
line-length = 88
unfixable = ["F401"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
